backend/requirements.txt
# Core Django requirements - Windows compatible
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
openai==1.3.0

# Optional packages (install separately if needed)
# Pillow  # For image handling - install separately: pip install Pillow
# django-extensions  # For development tools
# psycopg2-binary  # For PostgreSQL - use SQLite for development
# celery  # For background tasks
# redis  # For caching and Celery broker
